@@ -1,7 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Http, RequestOptionsArgs, URLSearchParams} from '@angular/http';
-import 'rxjs/add/operator/map'
-import 'rxjs/add/operator/toPromise';
+import {Http, URLSearchParams} from '@angular/http';
 import {Run} from "./run";
 import {environment} from "../../environments/environment";
 import {UtilsService} from "./utils.service";
